+++ src/main/java/org/example/datn_website_be/model/Promotion.java
@@ -33,7 +33,7 @@ public class Promotion extends BaseEntity {
     private String name;
 
     @Column
-    private double value;
+    private Integer value;
 
     @Column
     private Integer type;
